Implement the adding question method
Listen! the original number of questions was 30. So, if user wanted to
enter 6 question the size would increase to 36.So,i used the for loop to
 increase the size of ArrayList by 1 in each iteration so if n= 5
1+1+1+1+1 =5 this how loop will work.
 Now the method would work properly but their was still an issue that it
would always add the question at last index. The quiz game will
not be fun if user entered numbers are at the end. So, i did this i
used random class to generate a random number from 0 to size of
ArrayList (30 before 1st iteration, 31 in next and so on) added an
empty element in ArrayList , set the random number index element to
last index of list added the user entered question to that random number
index.
ez right
And the reason i used string cat in my options is to
display them in same format as my original options.